fix some warnings

+++ artnet-mini/artnet.h
@@ -2,13 +2,12 @@
 #define _ARTNET_H 1
 
 #include "artnet_node.h"
-#include <sys/types.h>
 
 #define ARTNET_PORT 6454
 
 typedef struct artnet_packet_s {
 	ipv4_addr_t ip;
-	size_t len;
+	uint16_t len;
 	uint8_t *data;
 } artnet_packet_t;
